% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diminui.R
\name{diminui}
\alias{diminui}
\title{Função diminui}
\usage{
diminui(a, b)
}
\arguments{
\item{b}{}
}
\description{
Esta é uma função simples para diminuir dois números
}
\examples{
diminui(3,2)
diminui(5,3)

}
