% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soma.R
\name{soma}
\alias{soma}
\title{Função soma}
\usage{
soma(a, b)
}
\arguments{
\item{a}{Primeiro número}

\item{b}{Segundo número}
}
\description{
Esta função é simples e soma dois números
}
\examples{
soma(2,3)
soma(4,5)

}
